Copy data out of pending to avoid races
If a pending_write is invalidated/completed while we're checking for
time-outs in `_check_pending` we could get into a situation where the
time field is `None`.

To resolve this, copy data so that it can't change out from underneath
us and make sure that potentially racy method calls will succeeded even
if the recod is invalidated.

There may still be races here, but to alleviate this, make sure that
data in record is still in a state where operations on it's fields will
succeed.