Allow tracking view to age out old data [ESD-902] (#930)
(v2.2 version of #926)

Tracking view adds data to it's CN0_dict memory but never ages out old data, thus anything that shows up will forever be held in the dictionary, this meant that logic to delete plots and plot data was never triggered. Add a mechanism to figure out if data in the dictionary is older than the oldest currently active plot point and remove it if it is.