Use an expiration map to reduce overhead
Having a large window is intended to allow fileio to always have
available window so calls to _check_pending should mostly be limited to
the end of a file transfer.

However, we can reduce the overhead of scanning a large window for write
completion by using a map to check if any packets are due for expiration
at the current time.  Sent packets are given an entry into an
"expireation map" tracking the epoch time in the future at which it
expires.

The Time object introduced limits time to seconds plus milliseconds.  If
we enter _check_pending we do a query of the expiration map using the
range of times from the last check to the current time.

The goal being that most calls to _check_pending should result in
only N dictionary queries (where N is the number of milliseconds since
the last check).  If packets do start to expire, then hopefully they'll
show up in relatively small "buckets" of expired packets in the
map.