Tweak the "Refresh settings from device" button [ESD-1268]
The previously separate button text moved into a button label to be
similar with the other buttons.

There is a bug in enable's SVGButton that horizontal layout doesn't
work (probably the reason why this button was originally configured
separately for button and text). A separate PR to be made to fix that.